First pass on making OpenGL ES1 compatible. Changes in draw:: should be transparent to non-ES. In draw:: push/pop 2D/3D removed; use View3D instead.

+++ include/glv_view3D.h
@@ -24,13 +24,13 @@ public:
 	/// 3D drawing callback
 	virtual void onDraw3D(GLV& g){}
 
-	double far() const { return mFar; }			///< Get far clip distance
-	double near() const { return mNear; }		///< Get near clip distance
-	double fovy() const { return mFOVY; }		///< Get field of view angle, in degrees, in the y direction
+	float far() const { return mFar; }			///< Get far clip distance
+	float near() const { return mNear; }		///< Get near clip distance
+	float fovy() const { return mFOVY; }		///< Get field of view angle, in degrees, in the y direction
 
-	void far(double v){ mFar=v; }				///< Set far clip distance
-	void near(double v){ mNear=v; }				///< Set near clip distance
-	void fovy(double v){ mFOVY=v; }				///< Set field of view angle, in degrees, in the y direction
+	void far(float v){ mFar=v; }				///< Set far clip distance
+	void near(float v){ mNear=v; }				///< Set near clip distance
+	void fovy(float v){ mFOVY=v; }				///< Set field of view angle, in degrees, in the y direction
 
 	/// Set 4x4 model transform matrix (column-major)
 	template <class T>
@@ -39,12 +39,12 @@ public:
 	}
 	
 	/// Get 4x4 model transform matrix
-	double * matrix(){ return mMatrix; }
+	float * matrix(){ return mMatrix; }
 
 protected:
 	virtual void onDraw(GLV& g);
-	double mNear, mFar, mFOVY;
-	double mMatrix[16];
+	float mNear, mFar, mFOVY;
+	float mMatrix[16];
 	
 //	struct ViewModel : public Model{
 //		ViewModel(View3D& self): v(self){}
